ft_transcendence-main/src/game-web/requirements.txt
Django>=5.0,<6.0
psycopg2>=2.9,<3.0
pillow>=10.2.0
requests>=2.0
djangorestframework>=3.14,<3.15
djangorestframework-simplejwt>=5.3,<5.5
channels>=4.0.0,<4.1.0
daphne>=4.1.0,<4.2.0
channels-redis>=4.2.0,<4.3.0
